move the visible contract test into AbstractPruningTest and make Postgres and Oracle versions
