+++ sdk/ledger-service/http-json/src/it/edition/ee/com/daml/http/PostgresPruningTest.scala
@@ -1,0 +1,6 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.http
+
+final class PostgresPruningTest extends AbstractPruningTest with HttpServicePostgresInt {}
